Relocate and rename the user database file
