+++ server/db/queries/user.js
@@ -1,4 +1,4 @@
-const db = require('./connection');
+const db = require('../connection');
 
 const query = (text, params, callback) => {
   const start = Date.now();
